requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
azure-core==1.4.0
azure-storage-blob==12.3.1
black==19.10b0
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
coverage==5.1
cryptography==2.9.2
distlib==0.3.0
Django==3.0.5
django-cors-headers==3.2.1
Faker==4.0.3
filelock==3.0.12
identify==1.4.15
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
isodate==0.6.0
mixer==6.1.3
more-itertools==8.2.0
msrest==0.6.13
nodeenv==1.3.5
oauthlib==3.1.0
packaging==20.3
pathspec==0.8.0
pluggy==0.13.1
pre-commit==2.3.0
psycopg2==2.8.5
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2020.1
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.18
wcwidth==0.1.9
zipp==3.1.0
